# Slight perf improvement on char::to_ascii_lowercase